Reduce Technical Debt: Unit and Integration Tests (#9)
This PR reduces technical debt by improving the unit and integration tests in Whisper. 
The primary change required was to repackage some of the code and implement interfaces 
for mocks -- particularly mocks for Google Secret Manager, the only external service 
currently being used with Whisper. There are unit tests for the library code, and for both the 
server and the client, independently testing against a test server and test request respectively.