Deploy Whisper to Cloud Run with GitHub Actions
Some setup is required on Google and GitHub but it seems that the
simplest way to deploy both the React UI and the API is to use GitHub
actions to build gcr.io images and push them to Google, then run gcloud
to deploy to Cloud Run. This will stop us from using Google Build, but
hopefully that will make things simpler.